move anime grid into folder
